// app/api/db/schema.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Handicap {
  date: Timestamp;
  handicapIndex: number;
  matricula: number;
}

export interface Jugadores {
  clubName: string;
  fullName: string;
  matricula: number;
}

export interface DB {
  handicap: Handicap;
  jugadores: Jugadores;
}